Testability changes: - separated OnlineModeConfirmActions.js from OnlineModeActions.js - separated ImportOnlineSearchActions.js from ImportOnlineActions.js
